prototype-an-assistant.rst page, train button
